Change from rollup to webpack
Working around issues like "process is not defined" and
"module is not defined" with replaces, then stuff like having to
manually specify exports in commonjs files I can live with, but
when the new react rc totally broke the app and tests would not
run without that rc, that got the better of me. Back to familiar
waters with webpack.

/rant