Make Drag and drop work again
However, this means we have to give up the idea that the request
is a <button>. This is because the button will eat the dragstart
event if it is the recipient of the mousedown. Thus it is
impossible (AFAIK) to have a draggable wrapping a button and make
it work as intended.